% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{qda_plot_mds}
\alias{qda_plot_mds}
\title{Plot the code map}
\usage{
qda_plot_mds(fragments, unit = "annotationKey", min_n = 3)
}
\arguments{
\item{fragments}{A fragments data frame.}

\item{unit}{Column identifying the segment; defaults to `"annotationKey"`.}

\item{min_n}{Ignore codes used fewer than `min_n` times.  Rarely used
codes make every coefficient unstable.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot the code map
}
\examples{
frag <- qda_read_fragments(qda_example("easyqda-fragments.csv"))
qda_plot_mds(frag, min_n = 1)
}
