% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qda_plot_timeline}
\alias{qda_plot_timeline}
\title{Coding progress over time}
\usage{
qda_plot_timeline(history)
}
\arguments{
\item{history}{A history data frame from [qda_read_history()].}
}
\value{
A 'ggplot2' object.
}
\description{
The plugin's process view: how the number of codings grew, per coder.
}
\examples{
h <- qda_read_history(qda_example("zotqda-history.csv"))
qda_plot_timeline(h)
}
