% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{qda_read_codebook}
\alias{qda_read_codebook}
\title{Read the code system}
\usage{
qda_read_codebook(path)
}
\arguments{
\item{path}{Path to the CSV.}
}
\value{
A data frame; see [qda_read()].
}
\description{
Read the code system
}
\examples{
cb <- qda_read_codebook(qda_example("zotqda-codebook.csv"))
cb$code
}
