% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning.R
\name{qda_theme_power}
\alias{qda_theme_power}
\title{The same question as power}
\usage{
qda_theme_power(n, prevalence, instances = 1)
}
\arguments{
\item{n}{Number of documents.}

\item{prevalence}{Share of the population in which the theme is present.}

\item{instances}{How many separate occurrences you want to see.}
}
\value{
A probability.
}
\description{
Given the documents you have, how likely are you to meet a theme of this
prevalence the desired number of times?
}
\examples{
qda_theme_power(32, 0.05)          # about 0.8, the flip side of the table
}
