% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdaR-package.R
\docType{package}
\name{qdaR-package}
\alias{qdaR}
\alias{qdaR-package}
\title{qdaR: analyse qualitative coding exported from Zotero}
\description{
Reads the versioned exchange files written by the Zotero plugins zotQDA
and qdaZ, checks them against the contract they ship with, reproduces the
plugins' graphics with 'ggplot2', and adds the inferential statistics the
plugins deliberately leave out.
}
\details{
The division of labour is intentional: zotQDA writes the data, qdaZ
describes it, and inferential claims are made here, where the person
making them has to choose the test.

Start with [qda_read_fragments()] and [qda_example()].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zotqda.org}
  \item \url{https://qdar.zotqda.org/}
  \item \url{https://github.com/easyqda/qdaR}
  \item Report bugs at \url{https://github.com/easyqda/qdaR/issues}
}

}
\author{
\strong{Maintainer}: fre.ms \email{fre.ms@fre.ms}

Authors:
\itemize{
  \item fre.ms \email{fre.ms@fre.ms}
}

}
\keyword{internal}
